Adding authentication

+++ flash-review-backend/src/Database.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE LambdaCase          #-}
+
 {-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE RecordWildCards     #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -19,16 +19,22 @@ module Database
   , processReviewDb
   , getDueCountDb
   , setupSchema
+  , authenticateUserDb
+  , signupUserDb
   ) where
 
 import qualified API
 import           Control.Exception                  (SomeException, bracket,
                                                      throwIO, try)
 import           Control.Monad                      (void, when)
+import           Crypto.BCrypt                      (hashPasswordUsingPolicy,
+                                                     slowerBcryptHashingPolicy,
+                                                     validatePassword)
 import           Data.ByteString                    (ByteString)
 import qualified Data.ByteString.Char8              as BS
 import           Data.Maybe                         (fromMaybe)
 import           Data.Text                          (Text)
+import qualified Data.Text.Encoding                 as TE
 import           Data.Time                          (UTCTime, addUTCTime,
                                                      getCurrentTime)
 import           Data.UUID                          (UUID)
@@ -121,6 +127,7 @@ data FlashcardRow = FlashcardRow
   , rowInterval    :: Int
   , rowEaseFactor  :: Double
   , rowRepetitions :: Int
+  , rowUserId      :: UUID
   } deriving (Show)
 
 fromRow :: FlashcardRow -> API.Flashcard
@@ -132,6 +139,7 @@ fromRow FlashcardRow{..} = API.Flashcard
   , API.interval    = rowInterval
   , API.easeFactor  = rowEaseFactor
   , API.repetitions = rowRepetitions
+  , API.ownerId     = rowUserId
   }
 
 toRow :: API.Flashcard -> FlashcardRow
@@ -143,6 +151,7 @@ toRow card = FlashcardRow
   , rowInterval    = API.interval card
   , rowEaseFactor  = API.easeFactor card
   , rowRepetitions = API.repetitions card
+  , rowUserId      = API.ownerId card
   }
 
 instance PG.FromRow FlashcardRow where
@@ -153,7 +162,8 @@ instance PG.FromRow FlashcardRow where
     nextReview <- PG.field
     interval <- PG.field
     easeFactor <- PG.field
-    FlashcardRow uuid front back nextReview interval easeFactor <$> PG.field
+    repetitions <- PG.field
+    FlashcardRow uuid front back nextReview interval easeFactor repetitions <$> PG.field
 
 instance PG.ToRow FlashcardRow where
   toRow FlashcardRow{..} =
@@ -164,6 +174,46 @@ instance PG.ToRow FlashcardRow where
     , PG.toField rowInterval
     , PG.toField rowEaseFactor
     , PG.toField rowRepetitions
+    , PG.toField rowUserId
+    ]
+
+data UserRow = UserRow
+  { userId       :: UUID
+  , userName     :: Text
+  , userPassword :: Text
+  , userEmail    :: Text
+  } deriving (Show)
+
+fromUserRow :: UserRow -> API.User
+fromUserRow UserRow{..} = API.User
+  { API.userId = userId
+  , API.username = userName
+  , API.email = userEmail
+  , API.password = userPassword
+  }
+
+toUserRow :: API.User -> UserRow
+toUserRow API.User{..} = UserRow
+  { userId = userId
+  , userName = username
+  , userPassword = password
+  , userEmail = email
+  }
+
+instance PG.FromRow UserRow where
+  fromRow = do
+    userId <- PG.field
+    userName <- PG.field
+    userPassword <- PG.field
+    userEmail <- PG.field
+    return UserRow{..}
+
+instance PG.ToRow UserRow where
+  toRow UserRow{..} =
+    [ PG.toField userId
+    , PG.toField userName
+    , PG.toField userPassword
+    , PG.toField userEmail
     ]
 
 setupSchema :: PG.Connection -> IO ()
@@ -178,28 +228,69 @@ setupSchema conn = do
     \  next_review TIMESTAMP WITH TIME ZONE NOT NULL, \
     \  interval INTEGER NOT NULL, \
     \  ease_factor DOUBLE PRECISION NOT NULL, \
-    \  repetitions INTEGER NOT NULL \
+    \  repetitions INTEGER NOT NULL, \
+    \  user_id UUID NOT NULL \
+    \)"
+
+  void $ PG.execute_ conn
+    "CREATE TABLE IF NOT EXISTS users (\
+    \  userid UUID PRIMARY KEY, \
+    \  username TEXT NOT NULL UNIQUE, \
+    \  password TEXT NOT NULL, \
+    \  email TEXT NOT NULL UNIQUE \
     \)"
 
   void $ PG.execute_ conn
     "CREATE INDEX IF NOT EXISTS idx_flashcards_next_review ON flashcards (next_review)"
 
+  void $ PG.execute_ conn
+    "CREATE INDEX IF NOT EXISTS idx_flashcards_user_id ON flashcards (user_id)"
+
+  void $ PG.execute_ conn
+    "DO $$ \
+    \BEGIN \
+    \  IF NOT EXISTS ( \
+    \    SELECT 1 FROM information_schema.table_constraints \
+    \    WHERE constraint_name = 'fk_flashcards_user_id' \
+    \  ) THEN \
+    \    ALTER TABLE flashcards \
+    \    ADD CONSTRAINT fk_flashcards_user_id \
+    \    FOREIGN KEY (user_id) \
+    \    REFERENCES users(userid); \
+    \  END IF; \
+    \END; \
+    \$$"
+
+
+  void $ PG.execute_ conn
+    "DO $$ \
+    \BEGIN \
+    \  IF NOT EXISTS ( \
+    \    SELECT 1 FROM information_schema.columns \
+    \    WHERE table_name = 'flashcards' AND column_name = 'user_id' \
+    \  ) THEN \
+    \    ALTER TABLE flashcards ADD COLUMN user_id UUID; \
+    \  END IF; \
+    \END; \
+    \$$"
+
   putStrLn "Schema setup complete."
 
-getAllCardsDb :: PG.Connection -> IO [API.Flashcard]
-getAllCardsDb conn = do
-  rows <- PG.query_ conn
-    "SELECT id, front, back, next_review, interval, ease_factor, repetitions \
+getAllCardsDb :: PG.Connection -> UUID -> IO [API.Flashcard]
+getAllCardsDb conn userId = do
+  rows <- PG.query conn
+    "SELECT id, front, back, next_review, interval, ease_factor, repetitions, user_id \
     \FROM flashcards \
-    \ORDER BY next_review ASC"
+    \WHERE user_id = ? \
+    \ORDER BY next_review ASC" [userId]
   return $ map fromRow rows
 
-getCardByIdDb :: PG.Connection -> UUID -> IO (Maybe API.Flashcard)
-getCardByIdDb conn uuid = do
+getCardByIdDb :: PG.Connection -> UUID -> UUID -> IO (Maybe API.Flashcard)
+getCardByIdDb conn uuid userId = do
   rows <- PG.query conn
-    "SELECT id, front, back, next_review, interval, ease_factor, repetitions \
+    "SELECT id, front, back, next_review, interval, ease_factor, repetitions, user_id \
     \FROM flashcards \
-    \WHERE id = ?" [uuid]
+    \WHERE id = ? AND user_id = ?" [uuid, userId]
   return $ case rows of
     [row] -> Just (fromRow row)
     _     -> Nothing
@@ -208,8 +299,8 @@ createCardDb :: PG.Connection -> API.Flashcard -> IO API.Flashcard
 createCardDb conn card = do
   let row = toRow card
   void $ PG.execute conn
-    "INSERT INTO flashcards (id, front, back, next_review, interval, ease_factor, repetitions) \
-    \VALUES (?, ?, ?, ?, ?, ?, ?)" row
+    "INSERT INTO flashcards (id, front, back, next_review, interval, ease_factor, repetitions, user_id) \
+    \VALUES (?, ?, ?, ?, ?, ?, ?, ?)" row
   return card
 
 updateCardDb :: PG.Connection -> UUID -> API.Flashcard -> IO API.Flashcard
@@ -218,33 +309,33 @@ updateCardDb conn uuid card = do
   rowsAffected <- PG.execute conn
     "UPDATE flashcards \
     \SET front = ?, back = ?, next_review = ?, interval = ?, ease_factor = ?, repetitions = ? \
-    \WHERE id = ?"
+    \WHERE id = ? AND user_id = ?"
     (rowFront row, rowBack row, rowNextReview row, rowInterval row,
-     rowEaseFactor row, rowRepetitions row, uuid)
+     rowEaseFactor row, rowRepetitions row, uuid, rowUserId row)
 
   when (rowsAffected == 0) $ do
     void $ createCardDb conn card
 
   return card
 
-deleteCardDb :: PG.Connection -> UUID -> IO ()
-deleteCardDb conn uuid = do
+deleteCardDb :: PG.Connection -> UUID -> UUID -> IO ()
+deleteCardDb conn uuid userId = do
   void $ PG.execute conn
-    "DELETE FROM flashcards WHERE id = ?" [uuid]
+    "DELETE FROM flashcards WHERE id = ? AND user_id = ?" [uuid, userId]
 
-getReviewCardsDb :: PG.Connection -> IO [API.Flashcard]
-getReviewCardsDb conn = do
+getReviewCardsDb :: PG.Connection -> UUID -> IO [API.Flashcard]
+getReviewCardsDb conn userId = do
   now <- getCurrentTime
   rows <- PG.query conn
-    "SELECT id, front, back, next_review, interval, ease_factor, repetitions \
+    "SELECT id, front, back, next_review, interval, ease_factor, repetitions, user_id \
     \FROM flashcards \
-    \WHERE next_review <= ? \
-    \ORDER BY next_review ASC" [now]
+    \WHERE next_review <= ? AND user_id = ? \
+    \ORDER BY next_review ASC" (now, userId)
   return $ map fromRow rows
 
-processReviewDb :: PG.Connection -> UUID -> API.ReviewResult -> IO ()
-processReviewDb conn uuid result = do
-  maybeCard <- getCardByIdDb conn uuid
+processReviewDb :: PG.Connection -> UUID -> UUID -> API.ReviewResult -> IO ()
+processReviewDb conn uuid userId result = do
+  maybeCard <- getCardByIdDb conn uuid userId
   case maybeCard of
     Nothing -> pure ()
     Just card -> do
@@ -272,9 +363,40 @@ processReviewDb conn uuid result = do
         , API.nextReview = newNextReview
         }
 
-getDueCountDb :: PG.Connection -> IO Int
-getDueCountDb conn = do
+getDueCountDb :: PG.Connection -> UUID -> IO Int
+getDueCountDb conn userId = do
   now <- getCurrentTime
   [PG.Only count] <- PG.query conn
-    "SELECT COUNT(*) FROM flashcards WHERE next_review <= ?" [now]
+    "SELECT COUNT(*) FROM flashcards WHERE next_review <= ? AND user_id = ?" (now, userId)
   return count
+
+authenticateUserDb :: PG.Connection -> Text -> Text -> IO (Maybe API.User)
+authenticateUserDb conn username password = do
+  rows <- PG.query conn
+    "SELECT userid, username, password, email FROM users \
+    \WHERE username = ?" [username]
+  case rows of
+    [userRow@UserRow{..}] -> do
+
+      let storedHash = TE.encodeUtf8 userPassword
+          providedPass = TE.encodeUtf8 password
+      if validatePassword storedHash providedPass
+        then return $ Just (fromUserRow userRow)
+        else return Nothing
+    _ -> return Nothing
+
+signupUserDb :: PG.Connection -> API.User -> IO API.User
+signupUserDb conn user = do
+
+  let rawPassword = TE.encodeUtf8 $ API.password user
+  mHashedPass <- hashPasswordUsingPolicy slowerBcryptHashingPolicy rawPassword
+  case mHashedPass of
+    Nothing -> error "Failed to hash password"
+    Just hashedPass -> do
+
+      let hashedUser = user { API.password = TE.decodeUtf8 hashedPass }
+          UserRow{..} = toUserRow hashedUser
+      void $ PG.execute conn
+        "INSERT INTO users (userid, username, password, email) VALUES (?, ?, ?, ?)"
+        (userId, userName, userPassword, userEmail)
+      return hashedUser
